+++ packages/design-system-react-native/src/components/Button/variants/ButtonPrimary/ButtonPrimary.stories.tsx
@@ -2,7 +2,7 @@ import { useTailwind } from '@metamask/design-system-twrnc-preset';
 import type { Meta, StoryObj } from '@storybook/react-native';
 import { View } from 'react-native';
 
-import { IconName } from '../../../../Icon';
+import { IconName } from '../../../Icon';
 import { ButtonSize } from '../../Button.types';
 import ButtonPrimary from './ButtonPrimary';
 import { DEFAULT_BUTTONPRIMARY_PROPS } from './ButtonPrimary.constants';
