+++ packages/design-system-react-native/src/components/Button/variants/ButtonPrimary/ButtonPrimary.test.tsx
@@ -1,7 +1,7 @@
 import { render, fireEvent } from '@testing-library/react-native';
 import React from 'react';
 
-import { IconName } from '../../../../Icon';
+import { IconName } from '../../../Icon';
 import ButtonPrimary from './ButtonPrimary';
 import {
   generateButtonPrimaryContainerClassNames,
