+++ packages/design-system-react-native/src/components/Button/variants/ButtonPrimary/ButtonPrimary.tsx
@@ -1,37 +1,41 @@
-import { useTailwind } from '@metamask/design-system-twrnc-preset';
+import {
+  useTailwind,
+  withThemeProvider,
+  Theme,
+} from '@metamask/design-system-twrnc-preset';
 import React, { useMemo, useState } from 'react';
 import type { GestureResponderEvent } from 'react-native';
 
-import ButtonBase from '../../../../../primitives/ButtonBase';
-import type { SpinnerProps } from '../../../../../temp-components/Spinner';
-import type { IconProps, IconColor } from '../../../../Icon';
-import type { TextProps } from '../../../../Text/Text.types';
-import { DEFAULT_BUTTONTERTIARY_PROPS } from './ButtonTertiary.constants';
-import type { ButtonTertiaryProps } from './ButtonTertiary.types';
+import ButtonBase from '../../../../primitives/ButtonBase';
+import type { SpinnerProps } from '../../../../temp-components/Spinner';
+import type { IconProps, IconColor } from '../../../Icon';
+import type { TextProps } from '../../../Text/Text.types';
+import { DEFAULT_BUTTONPRIMARY_PROPS } from './ButtonPrimary.constants';
+import type { ButtonPrimaryProps } from './ButtonPrimary.types';
 import {
-  generateButtonTertiaryContainerClassNames,
-  generateButtonTertiaryTextClassNames,
-} from './ButtonTertiary.utilities';
+  generateButtonPrimaryContainerClassNames,
+  generateButtonPrimaryTextClassNames,
+} from './ButtonPrimary.utilities';
 
-const ButtonTertiary = ({
+const ButtonPrimaryBase = ({
   children,
   textProps,
   spinnerProps,
   startIconProps,
   endIconProps,
-  isDanger = DEFAULT_BUTTONTERTIARY_PROPS.isDanger,
-  isInverse = DEFAULT_BUTTONTERTIARY_PROPS.isInverse,
-  isLoading = DEFAULT_BUTTONTERTIARY_PROPS.isLoading,
+  isDanger = DEFAULT_BUTTONPRIMARY_PROPS.isDanger,
+  isInverse = DEFAULT_BUTTONPRIMARY_PROPS.isInverse,
+  isLoading = DEFAULT_BUTTONPRIMARY_PROPS.isLoading,
   onPressIn,
   onPressOut,
   twClassName,
   style,
   ...props
-}: ButtonTertiaryProps) => {
+}: ButtonPrimaryProps) => {
   const [isPressed, setIsPressed] = useState(false);
   const tw = useTailwind();
   const twContainerClassNames = useMemo(() => {
-    return generateButtonTertiaryContainerClassNames({
+    return generateButtonPrimaryContainerClassNames({
       isPressed,
       isDanger,
       isInverse,
@@ -41,7 +45,7 @@ const ButtonTertiary = ({
   }, [isPressed, isDanger, isInverse, isLoading, twClassName]);
 
   const twTextClassNames = useMemo(() => {
-    return generateButtonTertiaryTextClassNames({
+    return generateButtonPrimaryTextClassNames({
       isPressed,
       isDanger,
       isInverse,
@@ -50,30 +54,31 @@ const ButtonTertiary = ({
   }, [isPressed, isDanger, isInverse, isLoading]);
 
   const finalTextProps: Omit<Partial<TextProps>, 'children'> = {
-    ...DEFAULT_BUTTONTERTIARY_PROPS.textProps,
+    ...DEFAULT_BUTTONPRIMARY_PROPS.textProps,
     ...textProps,
     twClassName: `${twTextClassNames} ${textProps?.twClassName ?? ''}`,
   };
   const finalStartIconProps: Partial<IconProps> = {
-    ...DEFAULT_BUTTONTERTIARY_PROPS.startIconProps,
+    ...DEFAULT_BUTTONPRIMARY_PROPS.startIconProps,
     ...startIconProps,
     twClassName: `${twTextClassNames} ${startIconProps?.twClassName ?? ''}`,
   };
 
   const finalEndIconProps: Partial<IconProps> = {
-    ...DEFAULT_BUTTONTERTIARY_PROPS.endIconProps,
+    ...DEFAULT_BUTTONPRIMARY_PROPS.endIconProps,
     ...endIconProps,
-    twClassName: `${twTextClassNames} ${endIconProps?.twClassName ?? ''}`,
+    twClassName: `${twTextClassNames} ${startIconProps?.twClassName ?? ''}`,
   };
 
   const finalSpinnerProps: SpinnerProps = {
-    ...DEFAULT_BUTTONTERTIARY_PROPS.spinnerProps,
+    ...DEFAULT_BUTTONPRIMARY_PROPS.spinnerProps,
     color: twTextClassNames as IconColor,
     loadingTextProps: {
       twClassName: twTextClassNames,
     },
     ...spinnerProps,
   };
+
   const onPressInHandler = (event: GestureResponderEvent) => {
     setIsPressed(true);
     onPressIn?.(event);
@@ -94,7 +99,7 @@ const ButtonTertiary = ({
       onPressIn={onPressInHandler}
       onPressOut={onPressOutHandler}
       style={[tw`${twContainerClassNames}`, style]}
-      testID="button-tertiary"
+      testID="button-primary"
       {...props}
     >
       {children}
@@ -102,4 +107,15 @@ const ButtonTertiary = ({
   );
 };
 
-export default ButtonTertiary;
+const ButtonPrimaryLightOnly = withThemeProvider(
+  ButtonPrimaryBase,
+  Theme.Light,
+);
+
+const ButtonPrimary = ({ isInverse, ...props }: ButtonPrimaryProps) => {
+  if (isInverse) {
+    return <ButtonPrimaryBase isInverse={isInverse} {...props} />;
+  }
+  return <ButtonPrimaryLightOnly isInverse={isInverse} {...props} />;
+};
+export default ButtonPrimary;
