+++ packages/design-system-react-native/src/components/Button/variants/ButtonPrimary/ButtonPrimary.types.ts
@@ -1,4 +1,4 @@
-import type { ButtonBaseProps } from '../../../../../primitives/ButtonBase';
+import type { ButtonBaseProps } from '../../../../primitives/ButtonBase';
 
 /**
  * ButtonPrimary component props.
