+++ packages/design-system-react-native/src/components/Button/variants/ButtonSecondary/ButtonSecondary.tsx
@@ -2,10 +2,10 @@ import { useTailwind } from '@metamask/design-system-twrnc-preset';
 import React, { useMemo, useState } from 'react';
 import type { GestureResponderEvent } from 'react-native';
 
-import ButtonBase from '../../../../../primitives/ButtonBase';
-import type { SpinnerProps } from '../../../../../temp-components/Spinner';
-import type { IconProps, IconColor } from '../../../../Icon';
-import type { TextProps } from '../../../../Text/Text.types';
+import ButtonBase from '../../../../primitives/ButtonBase';
+import type { SpinnerProps } from '../../../../temp-components/Spinner';
+import type { IconProps, IconColor } from '../../../Icon';
+import type { TextProps } from '../../../Text/Text.types';
 import { DEFAULT_BUTTONSECONDARY_PROPS } from './ButtonSecondary.constants';
 import type { ButtonSecondaryProps } from './ButtonSecondary.types';
 import {
