+++ packages/design-system-react-native/src/components/Button/variants/ButtonTertiary/ButtonTertiary.utilities.ts
